Move `url` to our _extras_
